kble: move run function to app module
